{{cookiecutter.buildfolder}}/requirements.txt
cookiecutter
cognite-sdk
cognite-transformations-cli
click
python-dotenv
